Backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy

google-generativeai
pydantic
